+++ src/hixl/engine/hixl_options.h
@@ -38,7 +38,7 @@ struct ConnectPoolConfig {
 struct CommResourceConfigDesc {
   std::optional<std::vector<std::string>> protocol_desc;
   std::optional<uint32_t> listen_port;
-  std::optional<int32_t> qos;
+  std::optional<uint8_t> qos;
 };
 
 struct GlobalResourceConfig {
